+++ CHANGES.rst
@@ -1,7 +1,24 @@
 Changelog
 =========
 
-1.0a9 (unreleased)
+1.0a10 (unreleased)
+-------------------
+
+Bugfixes:
+
+- @search endpoint: Also prefill path query dict with context path.
+  This will allow users to supply an argument like path.depth=1, and still
+  have path.query be prefilled server-side to the context's path.
+  [lgraf]
+
+- Overhaul JSON schema generation for @types endpoint. It now returns
+  fields in correct order and in their appropriate fieldsets.
+  [lgraf]
+
+- Add missing id to the Plone site serialization, related to issue #186
+  [sneridagh]
+
+1.0a9 (2017-03-03)
 ------------------
 
 New Features: